refactor: extract `omnihook` crate for webhook notifications
