+++ src/action_dispatcher/webhook/mod.rs
@@ -1,12 +1,5 @@
 mod action;
-mod client;
 mod components;
-mod payload_builder;
 
 pub use action::WebhookAction;
-pub use client::{WebhookClient, WebhookConfig};
 pub use components::WebhookComponents;
-pub use payload_builder::{
-    DiscordPayloadBuilder, GenericWebhookPayloadBuilder, SlackPayloadBuilder,
-    TelegramPayloadBuilder, WebhookPayloadBuilder,
-};
